+++ plugins/org.eclipse.osee.orcs/src/org/eclipse/osee/orcs/transaction/TransactionFactory.java
@@ -47,7 +47,7 @@ public interface TransactionFactory {
 
    boolean setTxComment(TransactionId txId, String comment);
 
-   boolean replaceWithBaselineTxVersion(UserId userId, BranchId branchId, TransactionId txId, int artId, String comment);
+   boolean replaceWithBaselineTxVersion(UserId userId, BranchId branchId, TransactionId txId, Long artId, String comment);
 
    boolean purgeTxs(String txIds);
